+++ .vs/challenger_dev/ColorByRegexConfig.txt
@@ -1,0 +1,10 @@
+// Este arquivo contém regras para colorir guias de documentos por expressões regulares. Cada linha contém uma expressão regular que será testada em relação ao caminho completo de um arquivo. Todos os arquivos correspondentes a uma expressão regular compartilharão uma cor.
+// Você pode personalizar a cor atribuída a qualquer grupo de arquivos clicando com o botão direito do mouse na guia e escolhendo "Definir cor da guia".
+// As expressões regulares serão correspondidas na ordem em que aparecem neste arquivo. Veja https://docs.microsoft.com/en-us/dotnet/standard/base-types/regular-expressions para obter a sintaxe.
+// As expressões regulares não diferenciam maiúsculas de minúsculas por padrão. Você pode substituir esse comportamento usando opções de grupo de captura como "(?-i:expressão)".
+// Edite este arquivo e salve suas alterações para ver as alterações aplicadas imediatamente. Quaisquer erros encontrados durante a análise ou avaliação de expressões aparecerão na Janela de Saída em um painel chamado "Cor por Expressão Regular."
+^.*\.cs$
+^.*\.fs$
+^.*\.vb$
+^.*\.cp?p?$
+^.*\.hp?p?$
