+++ ForumAEVO/ForumAEVO/Models/DTOs/ComentarioDTO.cs
@@ -6,6 +6,7 @@
         public Guid UserId { get; set; }
         public Guid Id { get; set; }
         public string Msg { get; set; } =string.Empty;
+        public DateTime Data { get; set; }
     }
 
 }
