+++ ForumAEVO/ForumAEVO/Models/DTOs/TopicoDTO.cs
@@ -6,5 +6,6 @@
         public List<ComentarioDto>? Comentarios { get; set; }
         public Guid Id { get; set; }
         public string Msg { get; set; } = string.Empty;
+        public DateTime Data { get; set; }
     }
 }
